requirements.txt
# --- Core Frameworks ---
fastapi
uvicorn
python-dotenv
pydantic

# --- AI & Agents ---
langchain
langchain-groq
langgraph
langchain_google_genai
groq

# --- Financial Data & Search (CRITICAL for your Agents) ---
yfinance
pandas
pandas-ta        # Kept this! Your Technical Analyst likely needs it.
duckduckgo_search # Kept this! Your Risk Manager needs it.
beautifulsoup4   # For web scraping
lxml             # Helper for scraping

# --- Utilities ---
requests
httpx
numpy
colorama
tqdm
tenacity
typing_extensions